pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "mnist_cc"
version = "0.0.1"
description = "Mnist simple project using cookiecutter"
authors = [
  { name = "Lauge Rasmussen DTU", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
    "matplotlib>=3.10.8",
    "scikit-learn>=1.8.0",
    "snakeviz>=2.2.2",
    "tensorboard-plugin-profile>=2.21.3",
    "torch==2.6.0",
    "torch-tb-profiler>=0.4.3",
    "torchvision>=0.21.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.23.1",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "mypy>=1.19.1",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120
lint.select = ["I"]

[tool.coverage.run]
omit = ["tests/*"]
